requirements.txt
Django==1.11.5
Pillow==4.2.1
Unidecode==0.04.21
Willow==0.4
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
django-material==1.0.0
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.2
djangorestframework==3.6.4
html5lib==0.999999999
idna==2.6
olefile==0.44
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
wagtail==1.12.2
wagtailmenus==2.4.1
webencodings==0.5.1
wsgiref==0.1.2
